requirements.txt
requests~=2.32.0
PyYAML~=6.0.2
pytz~=2025.2
flask~=3.1.0
jsonpickle~=4.0.5
APScheduler~=3.11.0
numpy~=2.3.0